Home component added
