+++ client/src/App.js
@@ -1,40 +1,24 @@
-import React, {useState, useEffect} from 'react'
-import { Container, AppBar, Typography, Grow, Grid } from '@material-ui/core';
-import {useDispatch} from 'react-redux'
+import React from 'react'
+import { Container } from '@material-ui/core';
+import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
-import {getPosts} from './actions/posts';
-import Posts from './components/Posts/Posts';
-import Form from './components/Forms/Forms';
+import Navbar from './components/Navbar/Navbar';
+import Home from './components/Home/Home';
+import Auth from './components/Auth/Auth';
 
-import useStyles from './styles';
-import memories from './images/memories.png';
 const App = () => {
-    const [currentId, setCurrentId] = useState(0);
-    const classes = useStyles();
-    const dispatch = useDispatch();
 
-    useEffect(()=>{
-        dispatch(getPosts());
-    }, [currentId, dispatch])
     return (
-        <Container maxWidth="lg">
-            <AppBar className={classes.appBar} position="static" color="inherit">
-                <Typography className={classes.heading} variant="h2" align="center">Memories</Typography>
-                <img className={classes.image} src={memories} alt="icon" height="60" width="80" />
-            </AppBar>
-            <Grow in>
-                <Container>
-                    <Grid className={classes.mainContainer} container flexDirection="column-reverse" justifyContent="space-between" alignItems="stretch" spacing={3}>
-                        <Grid item xs={12} sm={7}>
-                            <Posts setCurrentId={setCurrentId}/>
-                        </Grid>
-                        <Grid item xs={12} sm={4}>
-                            <Form currentId={currentId} setCurrentId={setCurrentId}/>
-                        </Grid>
-                    </Grid>
-                </Container>
-            </Grow>
-        </Container>
+        <BrowserRouter>
+            <Container maxWidth="lg">
+                <Navbar />
+                <Switch>
+                    <Route path="/" exact component={Home} />
+                    <Route path="/auth" exact component={Auth} />
+                </Switch>
+                <Home />
+            </Container>
+        </BrowserRouter>
     )
 }
